requirements.txt
python-dotenv==1.0.1
langchain==0.2.5
langchain-community==0.2.5
unstructured==0.14.6
langchain-openai==0.1.8
langchain-pinecone==0.1.1